Server/requirements.txt
django
djangorestframework
psycopg2-binary
requests
django-filter